Fully relegated animation stuff to animation.lua
animation.lua now has an update callback. It takes in the data table for the current frame, and spits out an X offset, Y offset, and image to be drawn. The logic of timers and animation indexes is all fully contained within animation.lua, so all play_ghost.lua has to do is perform arithmetic on the coordinates and scroll offset.