Animation separation (#2)
* separated animation data to its own file

Those 230 lines of bullshit were really starting to clutter up play_ghost.lua. All that data is in animation.lua now, which returns a nice little module for you c:

The module also contains an update callback. It takes in the data table for the current frame, and spits out an X offset, Y offset, and image to be drawn. The logic of timers and animation indexes is all fully contained within animation.lua, so all play_ghost.lua has to do is perform arithmetic on the coordinates and scroll offset.