Eliminated horizontal wrapping!
An update to the ghost version was necessary to do this.
Ghost v2 no longer stores X scroll, since I realized that was redundant; instead, it stores the screen number. Using this value (which can be interpreted as the high byte of Mega Man's X position), we can determine whether the ghost is wrapping or truly ahead/behind with a simple formula:
drawX-drawXEmu == (screen*256+xPos) - (screenEmu*256+xPosEmu)
Proper draws will satisfy this formula. Improper wrapped draws will not.
Some hardcoded hacky checks for the screen scrolling state were necessary, since Rockman 2 sets the screen number a bit sooner than it really should when it does edge-of-the-screen scrolls.