implement ghost v4 (#5)
* ghosts are now multi-stage, synced to loading lag.

This means one ghost file can effectively store an entire speedrun!
Brief segments of gameplay can still be captured by starting in the middle of
the stage instead of during loading lag.

* ghosts can be hidden for up to 65,535 frames.

While an animIndex of 0xFF hides the ghost for the current frame, the data
packet still exists. This tells play_ghost to wait up to 65,535 frames to
prevent wasting all that space.

* drop support for v0-v2

That was getting annoying. Who else is using this code, anyway?

* gitignore all ghosts but samples

* fixup filedlg parameter names

I can't remember if these didn't work when all lowercase or what.